Code/API/Python/crawler/requirements.txt
beautifulsoup4==4.5.1
ez-setup==0.9
mysql-connector-python==2.1.3
requests==2.11.1
